requirements.txt
flask==2.3.2
requests
psycopg2-binary
